+++ src/prelude.rs
@@ -2,12 +2,9 @@
 // See the COPYRIGHT file at the top-level directory of this distribution.
 // Licensed under the MIT license, see the LICENSE file or <http://opensource.org/licenses/MIT>
 
-pub use {
-    Cast,
-    Continue,
-    StaticType,
-    ToValue,
-};
+//! Traits and essential types inteded for blanket imports.
+
+pub use glib::prelude::*;
 
 #[cfg(feature = "3.4")]
 pub use auto::traits::ActionableExt;